Add files via upload
In order to discuss the most influential spreader, we treat the same set of data as separate 
undirected and directed graphs. In the part of the undirected graph, we delineated the communities 
by the greedy algorithm and the label propagation method, respectively, and compared the accuracy 
of the two methods for delineating communities by Q-value. In the undirected graph, we also 
measure the whole set of data by the average shortest path length. In the directed graph, we calculate 
and output the values of in-degree and out-degree to give the most influential nodes by comparing 
the degree of connectivity of particular points with other points over time through the preferential 
attachment model to derive the points with the largest degree that are more likely to get better in the 
development process and the DeGroot model compares the degree of connectivity of particular 
points with other points over time, thus giving the most influential propagators.